♻️ Refactor: HeadNav의 프로필 표시 관련 컴포넌트 분리
