+++ src/components/layout/HeadNav/HeadNav.tsx
@@ -1,48 +1,8 @@
-import { useEffect, useState } from "react";
 import * as S from "./HeadNav.style";
 import Link from "next/link";
 import Image from "next/image";
-import { useCurrentUser } from "@/src/context/UserContext";
 
-const ProfileData = function () {
-  const userData = useCurrentUser();
-  const loginStatus = !!userData;
-
-  const [accountEmail, setAccountEmail] = useState("");
-  const [profileImg, setProfileImg] = useState(
-    "/assets/icons/svg/nav-profile-default.svg"
-  );
-
-  useEffect(() => {
-    if (loginStatus) {
-      setAccountEmail(userData.email);
-      setProfileImg(userData.image_source);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return (
-    <>
-      {loginStatus && (
-        <S.NavProfileSection>
-          <Image
-            width={28}
-            height={28}
-            src={profileImg}
-            alt="loggedInProfileImg"
-          />
-          {accountEmail}
-        </S.NavProfileSection>
-      )}
-
-      {!loginStatus && (
-        <Link href={"/login"}>
-          <S.NavLoginButton>로그인</S.NavLoginButton>
-        </Link>
-      )}
-    </>
-  );
-};
+import HeadNavProfile from "./Profile/Profile";
 
 export default function HeadNav({ isSticky = true }) {
   return (
@@ -56,7 +16,7 @@ export default function HeadNav({ isSticky = true }) {
             alt="Linkbrary Logo"
           />
         </Link>
-        <ProfileData />
+        <HeadNavProfile />
       </S.NavWrapper>
     </S.Nav>
   );
